ChatingSystem/requirements.txt
amqp==5.3.1
asgiref==3.9.2
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.2
cachetools==6.2.0
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.2
daphne==4.2.1
Django==5.2.7
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
httplib2==0.31.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.4
Markdown==3.9
msgpack==1.1.1
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
zope.interface==8.0.1
